devtools: Split tests into modules per tab
